+++ R/adasyn.R
@@ -74,7 +74,6 @@
 #'   geom_point() +
 #'   labs(title = "With ADASYN")
 #'
-#' @importFrom recipes rand_id add_step ellipse_check
 step_adasyn <-
   function(recipe, ..., role = NA, trained = FALSE, column = NULL,
            over_ratio = 1, neighbors = 5, skip = TRUE,
@@ -88,16 +87,16 @@ step_adasyn <-
                column = column,
                over_ratio = over_ratio,
                neighbors = neighbors,
+               predictors = NULL,
                skip = skip,
                seed = seed,
                id = id
              ))
   }
 
-#' @importFrom recipes step
 step_adasyn_new <-
-  function(terms, role, trained, column, over_ratio, neighbors, skip, seed,
-           id) {
+  function(terms, role, trained, column, over_ratio, neighbors, predictors,
+           skip, seed, id) {
     step(
       subclass = "adasyn",
       terms = terms,
@@ -106,6 +105,7 @@ step_adasyn_new <-
       column = column,
       over_ratio = over_ratio,
       neighbors = neighbors,
+      predictors = predictors,
       skip = skip,
       id = id,
       seed = seed,
@@ -113,9 +113,6 @@ step_adasyn_new <-
     )
   }
 
-#' @importFrom recipes bake prep check_type
-#' @importFrom dplyr select
-#' @importFrom purrr map_lgl
 #' @export
 prep.step_adasyn <- function(x, training, info = NULL, ...) {
 
@@ -125,7 +122,9 @@ prep.step_adasyn <- function(x, training, info = NULL, ...) {
   if (!is.factor(training[[col_name]]))
     rlang::abort(paste0(col_name, " should be a factor variable."))
 
-  check_type(select(training, -col_name), TRUE)
+  predictors <- setdiff(info$variable[info$role == "predictor"], col_name)
+
+  check_type(training[, predictors], TRUE)
 
   if (any(map_lgl(training, ~ any(is.na(.x)))))
     rlang::abort("`NA` values are not allowed when using `step_adasyn`")
@@ -137,32 +136,36 @@ prep.step_adasyn <- function(x, training, info = NULL, ...) {
     column = col_name,
     over_ratio = x$over_ratio,
     neighbors = x$neighbors,
+    predictors = predictors,
     skip = x$skip,
     seed = x$seed,
     id = x$id
   )
 }
 
-#' @importFrom tibble as_tibble tibble
-#' @importFrom withr with_seed
-#' @importFrom dplyr mutate
-#' @importFrom rlang :=
 #' @export
 bake.step_adasyn <- function(object, new_data, ...) {
+
+  new_data <- as.data.frame(new_data)
+
+  predictor_data <- new_data[, unique(c(object$predictors, object$column))]
 
   # adasyn with seed for reproducibility
   with_seed(
     seed = object$seed,
     code = {
-      new_data <- adasyn(new_data, object$column,
-                         k = object$neighbors, over_ratio = object$over_ratio)
+      synthetic_data <- adasyn(
+        predictor_data,
+        object$column,
+        k = object$neighbors,
+        over_ratio = object$over_ratio)
     }
   )
+  new_data <- na_splice(new_data, synthetic_data, object)
 
   as_tibble(new_data)
 }
 
-#' @importFrom recipes printer terms_select
 #' @export
 print.step_adasyn <-
   function(x, width = max(20, options()$width - 26), ...) {
@@ -173,8 +176,6 @@ print.step_adasyn <-
 
 #' @rdname step_adasyn
 #' @param x A `step_adasyn` object.
-#' @importFrom generics tidy
-#' @importFrom recipes sel2char is_trained
 #' @export
 tidy.step_adasyn <- function(x, ...) {
   if (is_trained(x)) {