+++ R/bsmote.R
@@ -111,7 +111,6 @@
 #'   geom_point() +
 #'   labs(title = "With borderline-SMOTE, all_neighbors = TRUE")
 #'
-#' @importFrom recipes rand_id add_step ellipse_check
 step_bsmote <-
   function(recipe, ..., role = NA, trained = FALSE,
            column = NULL, over_ratio = 1, neighbors = 5, all_neighbors = FALSE,
@@ -126,16 +125,16 @@ step_bsmote <-
                over_ratio = over_ratio,
                neighbors = neighbors,
                all_neighbors = all_neighbors,
+               predictors = NULL,
                skip = skip,
                seed = seed,
                id = id
              ))
   }
 
-#' @importFrom recipes step
 step_bsmote_new <-
-  function(terms, role, trained, column, over_ratio, neighbors, all_neighbors, skip,
-           seed, id) {
+  function(terms, role, trained, column, over_ratio, neighbors, all_neighbors,
+           predictors, skip, seed, id) {
     step(
       subclass = "bsmote",
       terms = terms,
@@ -145,6 +144,7 @@ step_bsmote_new <-
       over_ratio = over_ratio,
       neighbors = neighbors,
       all_neighbors =  all_neighbors,
+      predictors = predictors,
       skip = skip,
       id = id,
       seed = seed,
@@ -152,8 +152,6 @@ step_bsmote_new <-
     )
   }
 
-#' @importFrom recipes bake prep check_type
-#' @importFrom dplyr select
 #' @export
 prep.step_bsmote <- function(x, training, info = NULL, ...) {
 
@@ -163,7 +161,9 @@ prep.step_bsmote <- function(x, training, info = NULL, ...) {
   if (!is.factor(training[[col_name]]))
     rlang::abort(paste0(col_name, " should be a factor variable."))
 
-  check_type(select(training, -col_name), TRUE)
+  predictors <- setdiff(info$variable[info$role == "predictor"], col_name)
+
+  check_type(training[, predictors], TRUE)
   check_na(select(training, -col_name), "step_bsmote")
 
   step_bsmote_new(
@@ -174,32 +174,36 @@ prep.step_bsmote <- function(x, training, info = NULL, ...) {
     over_ratio = x$over_ratio,
     neighbors = x$neighbors,
     all_neighbors = x$all_neighbors,
+    predictors = predictors,
     skip = x$skip,
     seed = x$seed,
     id = x$id
   )
 }
 
-#' @importFrom tibble as_tibble tibble
-#' @importFrom withr with_seed
 #' @export
 bake.step_bsmote <- function(object, new_data, ...) {
 
   new_data <- as.data.frame(new_data)
+
+  predictor_data <- new_data[, unique(c(object$predictors, object$column))]
   # bsmote with seed for reproducibility
   with_seed(
     seed = object$seed,
     code = {
-      new_data <- bsmote(new_data, object$column,
-                         k = object$neighbors, over_ratio = object$over_ratio,
-                         all_neighbors = object$all_neighbors)
+      synthetic_data <- bsmote(
+        predictor_data,
+        object$column,
+        k = object$neighbors,
+        over_ratio = object$over_ratio,
+        all_neighbors = object$all_neighbors)
     }
   )
+  new_data <- na_splice(new_data, synthetic_data, object)
 
   as_tibble(new_data)
 }
 
-#' @importFrom recipes printer terms_select
 #' @export
 print.step_bsmote <-
   function(x, width = max(20, options()$width - 26), ...) {
@@ -210,8 +214,6 @@ print.step_bsmote <-
 
 #' @rdname step_bsmote
 #' @param x A `step_bsmote` object.
-#' @importFrom generics tidy
-#' @importFrom recipes sel2char is_trained
 #' @export
 tidy.step_bsmote <- function(x, ...) {
   if (is_trained(x)) {