+++ R/rose.R
@@ -97,7 +97,6 @@
 #'   geom_point() +
 #'   labs(title = "With ROSE")
 #'
-#' @importFrom recipes rand_id add_step ellipse_check
 step_rose <-
   function(recipe, ..., role = NA, trained = FALSE,
            column = NULL, over_ratio = 1, minority_prop = 0.5,
@@ -114,16 +113,17 @@ step_rose <-
                minority_prop = minority_prop,
                minority_smoothness = minority_smoothness,
                majority_smoothness = majority_smoothness,
+               predictors = NULL,
                skip = skip,
                seed = seed,
                id = id
              ))
   }
 
-#' @importFrom recipes step
 step_rose_new <-
   function(terms, role, trained, column, over_ratio, minority_prop,
-           minority_smoothness, majority_smoothness, skip, seed, id) {
+           minority_smoothness, majority_smoothness, predictors, skip, seed,
+           id) {
     step(
       subclass = "rose",
       terms = terms,
@@ -134,14 +134,13 @@ step_rose_new <-
       minority_prop = minority_prop,
       minority_smoothness = minority_smoothness,
       majority_smoothness = majority_smoothness,
+      predictors = predictors,
       skip = skip,
       seed = seed,
       id = id
     )
   }
 
-#' @importFrom recipes bake prep check_type
-#' @importFrom dplyr select
 #' @export
 prep.step_rose <- function(x, training, info = NULL, ...) {
 
@@ -152,7 +151,10 @@ prep.step_rose <- function(x, training, info = NULL, ...) {
     rlang::abort(paste0(col_name, " should be a factor variable."))
 
   check_2_levels_only(training, col_name)
-  check_type(select(training, -col_name), TRUE)
+
+  predictors <- setdiff(info$variable[info$role == "predictor"], col_name)
+
+  check_type(training[, predictors], TRUE)
 
   step_rose_new(
     terms = x$terms,
@@ -163,6 +165,7 @@ prep.step_rose <- function(x, training, info = NULL, ...) {
     minority_prop = x$minority_prop,
     minority_smoothness = x$minority_smoothness,
     majority_smoothness = x$majority_smoothness,
+    predictors = predictors,
     skip = x$skip,
     seed = x$seed,
     id = x$id
@@ -170,9 +173,6 @@ prep.step_rose <- function(x, training, info = NULL, ...) {
 }
 
 
-#' @importFrom tibble as_tibble tibble
-#' @importFrom withr with_seed
-#' @importFrom ROSE ROSE
 #' @export
 bake.step_rose <- function(object, new_data, ...) {
   if (any(is.na(new_data[[object$column]])))
@@ -181,23 +181,33 @@ bake.step_rose <- function(object, new_data, ...) {
     missing <- NULL
 
   new_data <- as.data.frame(new_data)
+
+  predictor_data <- new_data[, unique(c(object$predictors, object$column))]
+
   # rose with seed for reproducibility
-  majority_size <- max(table(new_data[[object$column]])) * 2
+  majority_size <- max(table(predictor_data[[object$column]])) * 2
   with_seed(
     seed = object$seed,
     code = {
-      new_data <- ROSE(string2formula(object$column), new_data,
-                       N = majority_size * object$over_ratio,
-                       p = object$minority_prop,
-                       hmult.majo = object$majority_smoothness,
-                       hmult.mino = object$minority_smoothness)$data
+      original_levels <- levels(predictor_data[[object$column]])
+      synthetic_data <- ROSE(
+        string2formula(object$column),
+        predictor_data,
+        N = majority_size * object$over_ratio,
+        p = object$minority_prop,
+        hmult.majo = object$majority_smoothness,
+        hmult.mino = object$minority_smoothness
+      )
+      synthetic_data <- synthetic_data$data
+      synthetic_data[[object$column]] <- factor(synthetic_data[[object$column]],
+                                          levels = original_levels)
     }
   )
+  new_data <- na_splice(new_data, synthetic_data, object)
 
   as_tibble(new_data)
 }
 
-#' @importFrom recipes printer terms_select
 #' @export
 print.step_rose <-
   function(x, width = max(20, options()$width - 26), ...) {
@@ -208,8 +218,6 @@ print.step_rose <-
 
 #' @rdname step_rose
 #' @param x A `step_rose` object.
-#' @importFrom generics tidy
-#' @importFrom recipes sel2char is_trained
 #' @export
 tidy.step_rose <- function(x, ...) {
   if (is_trained(x)) {