+++ R/smote.R
@@ -88,7 +88,6 @@
 #'   geom_point() +
 #'   labs(title = "With SMOTE")
 #'
-#' @importFrom recipes rand_id add_step ellipse_check
 step_smote <-
   function(recipe, ..., role = NA, trained = FALSE,
            column = NULL, over_ratio = 1, neighbors = 5,
@@ -102,16 +101,16 @@ step_smote <-
                column = column,
                over_ratio = over_ratio,
                neighbors = neighbors,
+               predictors = NULL,
                skip = skip,
                seed = seed,
                id = id
              ))
   }
 
-#' @importFrom recipes step
 step_smote_new <-
-  function(terms, role, trained, column, over_ratio, neighbors, skip,
-           seed, id) {
+  function(terms, role, trained, column, over_ratio, neighbors, predictors,
+           skip, seed, id) {
     step(
       subclass = "smote",
       terms = terms,
@@ -120,6 +119,7 @@ step_smote_new <-
       column = column,
       over_ratio = over_ratio,
       neighbors = neighbors,
+      predictors = predictors,
       skip = skip,
       id = id,
       seed = seed,
@@ -127,8 +127,6 @@ step_smote_new <-
     )
   }
 
-#' @importFrom recipes bake prep check_type
-#' @importFrom dplyr select
 #' @export
 prep.step_smote <- function(x, training, info = NULL, ...) {
 
@@ -138,8 +136,11 @@ prep.step_smote <- function(x, training, info = NULL, ...) {
   if (!is.factor(training[[col_name]]))
     rlang::abort(paste0(col_name, " should be a factor variable."))
 
-  check_type(select(training, -col_name), TRUE)
+  predictors <- setdiff(info$variable[info$role == "predictor"], col_name)
+
+  check_type(training[, predictors], TRUE)
   check_na(select(training, -col_name), "step_smote")
+
 
   step_smote_new(
     terms = x$terms,
@@ -148,31 +149,37 @@ prep.step_smote <- function(x, training, info = NULL, ...) {
     column = col_name,
     over_ratio = x$over_ratio,
     neighbors = x$neighbors,
+    predictors = predictors,
     skip = x$skip,
     seed = x$seed,
     id = x$id
   )
 }
 
-#' @importFrom tibble as_tibble tibble
-#' @importFrom withr with_seed
 #' @export
 bake.step_smote <- function(object, new_data, ...) {
 
   new_data <- as.data.frame(new_data)
+
+  predictor_data <- new_data[, unique(c(object$predictors, object$column))]
+
   # smote with seed for reproducibility
   with_seed(
     seed = object$seed,
     code = {
-      new_data <- smote(new_data, object$column,
-                        k = object$neighbors, over_ratio = object$over_ratio)
+      synthetic_data <- smote(
+        predictor_data,
+        object$column,
+        k = object$neighbors,
+        over_ratio = object$over_ratio
+      )
     }
   )
+  new_data <- na_splice(new_data, synthetic_data, object)
 
   as_tibble(new_data)
 }
 
-#' @importFrom recipes printer terms_select
 #' @export
 print.step_smote <-
   function(x, width = max(20, options()$width - 26), ...) {
@@ -183,8 +190,6 @@ print.step_smote <-
 
 #' @rdname step_smote
 #' @param x A `step_smote` object.
-#' @importFrom generics tidy
-#' @importFrom recipes sel2char is_trained
 #' @export
 tidy.step_smote <- function(x, ...) {
   if (is_trained(x)) {