+++ NEWS.md
@@ -1,3 +1,9 @@
+# themis 0.1.2
+
+* tuneable steps now properly work with tune package.
+* Steps now Retain original factor level ordering. (#22)
+* Oversampling steps now ignore non-predictor variables when appropiate. (#20)
+
 # themis 0.1.1
 
 * `step_smote()` now work regardless of order of classes. Thanks to @sebastien-foulle for point it out #14.